requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
requests==2.31.0
pytest==8.0.0
pydantic==2.11.7
langchain==0.3.27
pypdf==5.9.0
snscrape==0.7.0.20230622
certifi==2025.7.14
firecrawl-py==2.16.3